Work on tests
Factorize test initialization; rely on container
Fix age checks
Factorize tests for member creation and checks
