+++ tests/Galette/Core/tests/units/Links.php
@@ -36,7 +36,7 @@
 
 namespace Galette\Core\test\units;
 
-use atoum;
+use Galette\GaletteTestCase;
 
 /**
  * Password tests class
@@ -50,37 +50,12 @@ use atoum;
  * @link      http://galette.tuxfamily.org
  * @since     2020-03-15
  */
-class Links extends atoum
+class Links extends GaletteTestCase
 {
     //private $pass = null;
-    private $zdb;
-    private $i18n;
-    private $preferences;
-    private $login;
-    private $history;
-    private $seed = 95842355;
-    private $adh;
+    protected $seed = 95842355;
     private $links;
     private $ids = [];
-    private $contribs_ids = [];
-    private $members_fields;
-    private $contrib;
-
-    /**
-     * Set up tests
-     *
-     * @return void
-     */
-    public function setUp()
-    {
-        $this->zdb = new \Galette\Core\Db();
-        $ct = new \Galette\Entity\ContributionsTypes($this->zdb);
-        if (count($ct->getCompleteList()) === 0) {
-            //status are not yet instanciated.
-            $res = $ct->installInit();
-            $this->boolean($res)->isTrue();
-        }
-    }
 
     /**
      * Set up tests
@@ -91,38 +66,13 @@ class Links extends atoum
      */
     public function beforeTestMethod($testMethod)
     {
-        $this->zdb = new \Galette\Core\Db();
+        parent::beforeTestMethod($testMethod);
+        $this->initStatus();
+        $this->initContributionsTypes();
+
         $this->links = new \Galette\Core\Links($this->zdb, false);
-
-        $this->i18n = new \Galette\Core\I18n(
-            \Galette\Core\I18n::DEFAULT_LANG
-        );
-
-        $this->preferences = new \Galette\Core\Preferences(
-            $this->zdb
-        );
-        $this->login = new \Galette\Core\Login($this->zdb, $this->i18n);
-        $this->history = new \Galette\Core\History($this->zdb, $this->login, $this->preferences);
-
-        global $zdb, $login, $hist, $i18n; // globals :(
-        $zdb = $this->zdb;
-        $login = $this->login;
-        $hist = $this->history;
-        $i18n = $this->i18n;
-
-        $status = new \Galette\Entity\Status($this->zdb);
-        if (count($status->getList()) === 0) {
-            $res = $status->installInit();
-            if ($res instanceof \Exception) {
-                throw $res;
-            }
-            $this->boolean($res)->isTrue();
-        }
-
         $this->contrib = new \Galette\Entity\Contribution($this->zdb, $this->login);
 
-        include_once GALETTE_ROOT . 'includes/fields_defs/members_fields.php';
-        $this->members_fields = $members_fields;
         $this->adh = new \Galette\Entity\Adherent($this->zdb);
         $this->adh->setDependencies(
             $this->preferences,
@@ -153,178 +103,6 @@ class Links extends atoum
     }
 
     /**
-     * Create test user in database
-     *
-     * @return void
-     */
-    private function createAdherent()
-    {
-        $bdate = new \DateTime(date('Y') . '-09-13');
-        //member is expected to be 28 years old
-        $bdate->sub(new \DateInterval('P28Y'));
-
-        $data = [
-            'nom_adh' => 'Hoarau',
-            'prenom_adh' => 'Lucas',
-            'ville_adh' => 'Reynaudnec',
-            'cp_adh' => '63077',
-            'adresse_adh' => '2, boulevard Legros',
-            'email_adh' => 'phoarau@tele2.fr',
-            'login_adh' => 'nathalie51',
-            'mdp_adh' => 'T.u!IbKOi|06',
-            'mdp_adh2' => 'T.u!IbKOi|06',
-            'bool_admin_adh' => false,
-            'bool_exempt_adh' => false,
-            'bool_display_info' => false,
-            'sexe_adh' => 1,
-            'prof_adh' => 'Extraction',
-            'titre_adh' => null,
-            'ddn_adh' => $bdate->format('Y-m-d'),
-            'lieu_naissance' => 'Fischer',
-            'pseudo_adh' => 'vallet.camille',
-            'pays_adh' => null,
-            'tel_adh' => '05 59 53 59 43',
-            'url_adh' => 'http://bodin.net/omnis-ratione-sint-dolorem-architecto',
-            'activite_adh' => true,
-            'id_statut' => 9,
-            'date_crea_adh' => '2019-05-20',
-            'pref_lang' => 'ca',
-            'fingerprint' => 'FAKER' . $this->seed,
-            'societe_adh' => 'Philippe',
-            'is_company' => true,
-        ];
-        $this->createMember($data);
-        $this->checkMemberExpected();
-    }
-
-    /**
-     * Create member from data
-     *
-     * @param array $data Data to use to create member
-     *
-     * @return \Galette\Entity\Adherent
-     */
-    public function createMember(array $data)
-    {
-        $adh = $this->adh;
-        $check = $adh->check($data, [], []);
-        if (is_array($check)) {
-            var_dump($check);
-        }
-        $this->boolean($check)->isTrue();
-
-        $store = $adh->store();
-        $this->boolean($store)->isTrue();
-
-        $this->ids[] = $adh->id;
-    }
-
-    /**
-     * Check members expecteds
-     *
-     * @param Adherent $adh           Member instance, if any
-     * @param array    $new_expecteds Changes on expected values
-     *
-     * @return void
-     */
-    private function checkMemberExpected($adh = null, $new_expecteds = [])
-    {
-        if ($adh === null) {
-            $adh = $this->adh;
-        }
-
-        $expecteds = [
-            'nom_adh' => 'Hoarau',
-            'prenom_adh' => 'Lucas',
-            'ville_adh' => 'Reynaudnec',
-            'cp_adh' => '63077',
-            'adresse_adh' => '2, boulevard Legros',
-            'email_adh' => 'phoarau@tele2.fr',
-            'login_adh' => 'nathalie51',
-            'mdp_adh' => 'T.u!IbKOi|06',
-            'bool_admin_adh' => false,
-            'bool_exempt_adh' => false,
-            'bool_display_info' => false,
-            'sexe_adh' => 1,
-            'prof_adh' => 'Extraction',
-            'titre_adh' => null,
-            'ddn_adh' => 'NOT USED',
-            'lieu_naissance' => 'Fischer',
-            'pseudo_adh' => 'vallet.camille',
-            'pays_adh' => '',
-            'tel_adh' => '05 59 53 59 43',
-            'url_adh' => 'http://bodin.net/omnis-ratione-sint-dolorem-architecto',
-            'activite_adh' => true,
-            'id_statut' => 9,
-            'pref_lang' => 'ca',
-            'fingerprint' => 'FAKER' . $this->seed,
-            'societe_adh' => 'Philippe'
-        ];
-        $expecteds = array_merge($expecteds, $new_expecteds);
-
-        foreach ($expecteds as $key => $value) {
-            $property = $this->members_fields[$key]['propname'];
-            switch ($key) {
-                case 'bool_admin_adh':
-                    $this->boolean($adh->isAdmin())->isIdenticalTo($value);
-                    break;
-                case 'bool_exempt_adh':
-                    $this->boolean($adh->isDueFree())->isIdenticalTo($value);
-                    break;
-                case 'bool_display_info':
-                    $this->boolean($adh->appearsInMembersList())->isIdenticalTo($value);
-                    break;
-                case 'activite_adh':
-                    $this->boolean($adh->isActive())->isIdenticalTo($value);
-                    break;
-                case 'mdp_adh':
-                    $pw_checked = password_verify($value, $adh->password);
-                    $this->boolean($pw_checked)->isTrue();
-                    break;
-                case 'ddn_adh':
-                    //rely on age, not on birthdate
-                    $this->variable($adh->$property)->isNotNull();
-                    $this->string($adh->getAge())->isIdenticalTo(' (28 years old)');
-                    break;
-                default:
-                    $this->variable($adh->$property)->isIdenticalTo(
-                        $value,
-                        "$property expected {$value} got {$adh->$property}"
-                    );
-                    break;
-            }
-        }
-
-        $d = \DateTime::createFromFormat('Y-m-d', $expecteds['ddn_adh']);
-
-        $expected_str = ' (28 years old)';
-        $this->string($adh->getAge())->isIdenticalTo($expected_str);
-        $this->boolean($adh->hasChildren())->isFalse();
-        $this->boolean($adh->hasParent())->isFalse();
-        $this->boolean($adh->hasPicture())->isFalse();
-
-        $this->string($adh->sadmin)->isIdenticalTo('No');
-        $this->string($adh->sdue_free)->isIdenticalTo('No');
-        $this->string($adh->sappears_in_list)->isIdenticalTo('No');
-        $this->string($adh->sstaff)->isIdenticalTo('No');
-        $this->string($adh->sactive)->isIdenticalTo('Active');
-        $this->variable($adh->stitle)->isNull();
-        $this->string($adh->sstatus)->isIdenticalTo('Non-member');
-        $this->string($adh->sfullname)->isIdenticalTo('HOARAU Lucas');
-        $this->string($adh->saddress)->isIdenticalTo('2, boulevard Legros');
-        $this->string($adh->sname)->isIdenticalTo('HOARAU Lucas');
-
-        $this->string($adh->getAddress())->isIdenticalTo($expecteds['adresse_adh']);
-        $this->string($adh->getAddressContinuation())->isEmpty();
-        $this->string($adh->getZipcode())->isIdenticalTo($expecteds['cp_adh']);
-        $this->string($adh->getTown())->isIdenticalTo($expecteds['ville_adh']);
-        $this->string($adh->getCountry())->isIdenticalTo($expecteds['pays_adh']);
-
-        $this->string($adh::getSName($this->zdb, $adh->id))->isIdenticalTo('HOARAU Lucas');
-        $this->string($adh->getRowClass())->isIdenticalTo('active cotis-never');
-    }
-
-    /**
      * Test new Link generation
      *
      * @return void
@@ -332,8 +110,8 @@ class Links extends atoum
     public function testGenerateNewLink()
     {
         $links = $this->links;
-        $this->createAdherent();
-        $id = current($this->ids);
+        $this->getMemberTwo();
+        $id = $this->adh->id;
 
         $res = $links->generateNewLink(
             \Galette\Core\Links::TARGET_MEMBERCARD,
@@ -355,7 +133,7 @@ class Links extends atoum
         $this->boolean($links->isHashValid(base64_encode('sthingthatisnotahash'), 'phoarau@tele2.fr'))->isFalse();
 
         $this->createContribution();
-        $cid = current($this->contribs_ids);
+        $cid = $this->contrib->id;
         $res = $links->generateNewLink(
             \Galette\Core\Links::TARGET_INVOICE,
             $cid
@@ -376,8 +154,8 @@ class Links extends atoum
     public function testExpiredValidate()
     {
         $links = $this->links;
-        $this->createAdherent();
-        $id = current($this->ids);
+        $this->getMemberTwo();
+        $id = $this->adh->id;
 
         $res = $links->generateNewLink(
             \Galette\Core\Links::TARGET_MEMBERCARD,
@@ -519,28 +297,6 @@ class Links extends atoum
         ];
         $this->createContrib($data);
         $this->checkContribExpected();
-    }
-
-    /**
-     * Create contribution from data
-     *
-     * @param array $data Data to use to create contribution
-     *
-     * @return \Galette\Entity\Contribution
-     */
-    public function createContrib(array $data)
-    {
-        $contrib = $this->contrib;
-        $check = $contrib->check($data, [], []);
-        if (is_array($check)) {
-            var_dump($check);
-        }
-        $this->boolean($check)->isTrue();
-
-        $store = $contrib->store();
-        $this->boolean($store)->isTrue();
-
-        $this->contribs_ids[] = (int)$contrib->id;
     }
 
     /**
